gui/requirements.txt
docstring_parser
pyside6
Pillow
numpy
matplotlib
# numba
